blackbox_backend/requirements.txt
Django
djangorestframework
django-phonenumber-field
django-phonenumbers
django-safedelete
django-cors-headers
mysql-connector-python
mysqlclient

